Added file structure
