--- src/App.js
@@ -1,29 +1,0 @@
-import React from 'react';
-// import logo from './logo.svg';
-import './App.css';
-import { makeStyles } from '@material-ui/core/styles';
-import Button from '@material-ui/core/Button';
-import Dashboard from './dashboard';
-
-const useStyles = makeStyles(theme => ({
-  button: {
-    margin: theme.spacing(1),
-  },
-  input: {
-    display: 'none',
-  },
-}));
-
-function App() {
-  const classes = useStyles();
-  return (
-    <div className="App">
-     <Dashboard></Dashboard>
-      
-    </div>
-
-    
-  );
-}
-
-export default App;
